+++ Samples/VideoStreamer/NetworkMediaStreamer.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29519.181
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "VideoStreamer", "VideoStreamerLib\build\VideoStreamer.vcxproj", "{58E4848E-B347-43FB-94C1-C5C85ABFD6B0}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NetworkMediaStreamer", "VideoStreamerLib\build\VideoStreamer.vcxproj", "{58E4848E-B347-43FB-94C1-C5C85ABFD6B0}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "VideoStreamerApp", "VideoStreamerApp\build\VideoStreamerApp.vcxproj", "{61D9C2A4-F52D-44DF-A262-7C5C8E7BB850}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NetworkMediaStreamerApp", "VideoStreamerApp\build\VideoStreamerApp.vcxproj", "{61D9C2A4-F52D-44DF-A262-7C5C8E7BB850}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
