+++ Samples/VideoStreamer/VideoStreamerLib/inc/RTSPServer.h
@@ -2,7 +2,6 @@
 #pragma once
 #define DBGLEVEL 1
 #include "RtspSession.h"
-#include <inspectable.h>
 
 class RTSPServer : public IRTSPServerControl
 {
